backend/requirements.txt
aiohttp==3.7.4
dnspython==1.16.0
mongoengine==0.20.0
pymongo==3.10.1
python-socketio==5.0.4
watchgod==0.6
